refactor: move graph tests to new graphs package
